prov/efa: move struct rxr_ep and related fucntion declarations to rxr_ep.h
Use a separate header file rxr_ep.h to store struct rxr_ep
and its related functions.
