+++ prov/efa/Makefile.include
@@ -90,6 +90,7 @@ _efa_headers = \
 	prov/efa/src/rdm/efa_rdm_peer.h	\
 	prov/efa/src/rdm/efa_rdm_cq.h \
 	prov/efa/src/rdm/rxr.h \
+	prov/efa/src/rdm/rxr_ep.h \
 	prov/efa/src/rdm/rxr_env.h \
 	prov/efa/src/rdm/rxr_cntr.h \
 	prov/efa/src/rdm/rxr_rma.h \
